ZipCode Validation failing
Valid zip codes failing validation. 

Added a '!' to TryParse to fail validation on invalid numbers.

The ShippingAddressInfo.ZipCode.Text InputScope is 'number' type. When input 'aaa', the 1st level of validation fails before the 'TryParse' is called.

There is no TextLable/Resource text linked with the ZipCode if the content is not a number. I'm new to Prism so will eventually work out what to change and add it to a PR.